+++ src/main/java/com/github/jlangch/venice/impl/docgen/DocGenerator.java
@@ -361,6 +361,7 @@ public class DocGenerator {
 		vectors.addSection(vec_create);
 		vec_create.addItem(getDocItem("[]"));
 		vec_create.addItem(getDocItem("vector"));
+		vec_create.addItem(getDocItem("vec"));
 		vec_create.addItem(getDocItem("mapv"));
 
 		final DocSection vec_access = new DocSection("Access");