+++ src/main/java/com/github/jlangch/venice/impl/types/collections/VncOrderedMap.java
@@ -113,6 +113,15 @@ public class VncOrderedMap extends VncMap {
 	}
 	
 	@Override
+	public VncVector toVncVector() {
+		return new VncVector(value
+							.entrySet()
+							.stream()
+							.map(e -> new VncVector(e.getKey(), e.getValue()))
+							.collect(Collectors.toList()));
+	}
+	
+	@Override
 	public int size() {
 		return value.size();
 	}