+++ ChangeLog.md
@@ -13,11 +13,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 - added function assoc-in
 - added function get-in
+- added function vec
 - added associative destructuring 
     - `(let [{:keys [a b]} {:a 1 :b 2}] (+ a b))`
     - `(let [{:syms [a b]} {'a 1 'b 2}] (+ a b))`
     - `(let [{:strs [a b]} {"a" 1 "b" 2}] (+ a b))`
     - `(fn [x {:keys [a b]}] (+ x a b))`
+- improved Java interop with Java functions returning arrays
+    - `byte[]` are converted to bytebuf
+    - all other arrays are converted to vector